Feature #67 Refactor ConnectThread::InitHostInfo to extract member variable initializations
Description

move to a standalone method and add to state transitions:
    memset(&host_info, 0, sizeof host_info);

    host_info.ai_family = AF_UNSPEC;
    host_info.ai_socktype = SOCK_STREAM; // Use SOCK_STREAM for TCP or SOCK_DGRAM for UDP.
    host_info.ai_flags = AI_PASSIVE;     // IP Wildcard

    if( host_info_list!=NULL )
    {
        freeaddrinfo(host_info_list);
        host_info_list = NULL;
    }

----------

extracted above to a separate method which took the name InitHostInfo, while the code that
    remained in the original function was renamed to the more descriptive InitAddressInfo.

This is the initial Checkin of code passing first compilation, but the target code is still untestable because of the way the ConnectThread starts up its thread in the constructor. Feature #89 has been opened to refactor the thread startups to standalone methods and tests can then be added for this feature, #67.