Task #87 - Refactor ConnectThread::ConnectThread to extract thread start to standalone method
extract

    ConnectionListenerThread = std::thread([=] { ConnectionListener(); });

to a ConnectThread::Run() for testability andmore refined control.

Update :

Instead of "Run" and "Stop" we'll call these methods "Start" and "Stop" because they are more
    conceptually symmetrical. (will change the other related tasks as well)

---------------------------------------------------------------------------------------------

Moved above listed code to start method. moved code from ~ConnectThread to Stop method.

Wrote tests for thread start and stop, and SocketState initial start and stop values.

Corrected issue with build of ..._unittest executable in Makefile that was preventing tests from running.

Tests are all green and code is clean, submitting for Feature Team review.

forgot to add Start and Stop calls to HarikanDawnServer for ConnectThread. corrected that.