Task 69 - Refactor ConnectThread::InitReceiveSocket to handle error conditions
added a stub to FlowControlModule.h & .c for LogMessage + the #defined log levels.

Added new SocketStates to support error handling operations including :
    - suspend and wait for console intervention
    - terminate server

NOTE : THIS PUSH IS NOT FINAL, but merely a share of a PROOF OF CONCEPT for other devs to see.
More pushes will replace this later with complete code.